requirements.txt
matplotlib==3.4.3
numpy==1.21.2
blake3==0.2.1
vaex-core==4.7
vaex-hdf5==0.11.1
pandas==1.3.3